+++ src/main/java/org/swyp/dessertbee/community/mate/controller/MateController.java
@@ -12,6 +12,7 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.swyp.dessertbee.common.annotation.ApiErrorResponses;
 import org.swyp.dessertbee.common.exception.BusinessException;
 import org.swyp.dessertbee.common.exception.ErrorCode;
 import org.swyp.dessertbee.community.mate.dto.request.MateCreateRequest;
@@ -19,9 +20,9 @@ import org.swyp.dessertbee.community.mate.dto.request.MateReportRequest;
 import org.swyp.dessertbee.community.mate.dto.response.MateDetailResponse;
 import org.swyp.dessertbee.community.mate.dto.response.MatesPageResponse;
 import org.swyp.dessertbee.community.mate.service.MateService;
+import org.swyp.dessertbee.community.mate.service.MateServiceImpl;
 import org.swyp.dessertbee.user.service.UserServiceImpl;
 
-import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
@@ -32,9 +33,9 @@ import java.util.UUID;
 @RestController
 @RequestMapping("api/mates")
 @RequiredArgsConstructor
-public class MateController {
+public class MateController{
 
-    private final MateService mateService;
+    private final MateServiceImpl mateService;
     private final UserServiceImpl userService;
 
     /**
@@ -45,6 +46,7 @@ public class MateController {
             @ApiResponse(responseCode = "200", description = "디저트메이트 생성 성공"),
             @ApiResponse(responseCode = "400", description = "잘못된 요청")
     })
+    @ApiErrorResponses({ErrorCode.USER_NOT_FOUND})
     @PostMapping(consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<MateDetailResponse> createMate(@RequestPart("request")  MateCreateRequest request,
                                                          @RequestPart(value = "mateImage", required = false) MultipartFile mateImage) {
@@ -59,7 +61,11 @@ public class MateController {
      * 메이트 상세 정보 조회
      */
     @Operation(summary = "메이트 상세 정보 조회", description = "디저트메이트 상세 정보 조회합니다.")
-//    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_USER', 'ROLE_ADMIN')")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트메이트 상세 정보 요청 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_NOT_FOUND})
     @GetMapping("/{mateUuid}")
     public ResponseEntity<MateDetailResponse> getMateDetail(@PathVariable UUID mateUuid) {
 
@@ -71,8 +77,13 @@ public class MateController {
     /**
      * 메이트 삭제
      */
-    @DeleteMapping("/{mateUuid}")
     @Operation(summary = "메이트 삭제", description = "디저트메이트 삭제합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트메이트 삭제 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_NOT_FOUND})
+    @DeleteMapping("/{mateUuid}")
     public ResponseEntity<Map<String, String>> deleteMate(@PathVariable UUID mateUuid) {
         //디저트메이트 삭제
         mateService.deleteMate(mateUuid);
@@ -86,8 +97,13 @@ public class MateController {
     /**
      * 메이트 수정
      * */
-    @PatchMapping("/{mateUuid}")
     @Operation(summary = "메이트 수정", description = "디저트메이트 수정합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트메이트 수정 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_NOT_FOUND})
+    @PatchMapping("/{mateUuid}")
     public ResponseEntity<Map<String, String>> updateMate(
             @PathVariable UUID mateUuid,
             @RequestPart(value = "request") MateCreateRequest request,
@@ -106,9 +122,13 @@ public class MateController {
     /**
      * 디저트메이트 전체 조회
      * */
-    @GetMapping
     @Operation(summary = "메이트 전체 조회", description = "디저트메이트 전체 조회합니다.")
-//    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_USER', 'ROLE_ADMIN')")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트메이트 전체 조회 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.INVALID_RANGE})
+    @GetMapping
     public ResponseEntity<MatesPageResponse> getMates(
             @RequestParam(required = false, defaultValue = "0") int from,
             @RequestParam(required = false, defaultValue = "10") int to,
@@ -133,8 +153,13 @@ public class MateController {
     /**
      * 내가 참여한 디저트메이트 조회
      * */
+    @Operation(summary = "내가 참여한 디저트메이트 조회", description = "내가 참여한 디저트메이트 조회합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "내가 참여한 디저트메이트 조회 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.INVALID_RANGE})
     @GetMapping("/me")
-//    @PreAuthorize("isAuthenticated() and hasAnyRole('ROLE_USER', 'ROLE_ADMIN')")
     public ResponseEntity<MatesPageResponse> getMyMates( @RequestParam(required = false, defaultValue = "0") int from,
                                                          @RequestParam(required = false, defaultValue = "10") int to){
 
@@ -151,6 +176,12 @@ public class MateController {
     /**
      * 디저트메이트 신고 기능
      * */
+    @Operation(summary = "디저트메이트 신고 기능", description = "디저트메이트 신고합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트메이트 신고 기능"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.DUPLICATION_REPORT})
     @PostMapping("/{mateUuid}/report")
     public ResponseEntity<Map<String, String>> reportMate(@PathVariable UUID mateUuid,
                                              @RequestBody  MateReportRequest request) {
