+++ src/main/java/org/swyp/dessertbee/community/mate/controller/MateMemberController.java
@@ -1,9 +1,14 @@
 package org.swyp.dessertbee.community.mate.controller;
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.swyp.dessertbee.common.annotation.ApiErrorResponses;
+import org.swyp.dessertbee.common.exception.ErrorCode;
 import org.swyp.dessertbee.community.mate.dto.request.MateApplyMemberRequest;
 import org.swyp.dessertbee.community.mate.dto.response.MateMemberResponse;
 import org.swyp.dessertbee.community.mate.service.MateMemberService;
@@ -24,6 +29,12 @@ public class MateMemberController {
     /**
      * 디저트 메이트 멤버 전체 조회
      * */
+    @Operation(summary = "디저트 메이트 멤버 전체 조회", description = "디저트 메이트 멤버 전체 조회합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 멤버 전체 조회 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.USER_NOT_FOUND})
     @GetMapping("/members")
     public ResponseEntity<List<MateMemberResponse>> getMembers(@PathVariable UUID mateUuid) {
 
@@ -35,6 +46,12 @@ public class MateMemberController {
     /**
      * 디저트 메이트 멤버 신청 api
      * */
+    @Operation(summary = "디저트 메이트 멤버 신청", description = "디저트 메이트 멤버 신청합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 멤버 신청 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_RECRUIT_DONE, ErrorCode.MATE_APPLY_BANNED, ErrorCode.MATE_APPLY_WAIT, ErrorCode.MATE_APPLY_REJECT, ErrorCode.ALREADY_TEAM_MEMBER})
     @PostMapping("/apply")
         public ResponseEntity<Map<String, String>> applyMate(@PathVariable UUID mateUuid) {
 
@@ -50,6 +67,12 @@ public class MateMemberController {
     /**
      * 디저트메이트 멤버 신청 취소 api
      * */
+    @Operation(summary = "디저트 메이트 멤버 신청 취소", description = "디저트 메이트 멤버 신청 취소합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 멤버 신청 취소 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_NOT_PENDING_MEMBER})
     @DeleteMapping("/apply")
     public ResponseEntity<Map<String, String>> cancelApplyMate(@PathVariable UUID mateUuid) {
 
@@ -60,9 +83,17 @@ public class MateMemberController {
         response.put("message", "디저트메이트 성공적으로 신청 취소되었습니다.");
         return ResponseEntity.ok(response);
     }
+
+
     /**
      * 디저트 메이트 대기 멤버 전체 조회
      **/
+    @Operation(summary = "디저트 메이트 대기 멤버 전체 조회", description = "디저트 메이트 대기 멤버 전체 조회합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 대기 멤버 전체 조회 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.USER_NOT_FOUND})
     @GetMapping("/pending")
     public ResponseEntity<List<MateMemberResponse>> pendingMate(@PathVariable UUID mateUuid) {
 
@@ -70,9 +101,17 @@ public class MateMemberController {
 
         return ResponseEntity.ok(members);
     }
+
+
     /**
      * 디저트 메이트 멤버 신청 수락 api
      * */
+    @Operation(summary = "디저트 메이트 멤버 신청 수락", description = "디저트 메이트 멤버 신청 수락합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 멤버 신청 수락 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_MEMBER_NOT_FOUND})
     @PatchMapping("/apply")
     public ResponseEntity<Map<String, String>> acceptMember(@PathVariable UUID mateUuid, @RequestBody MateApplyMemberRequest request) {
 
@@ -87,6 +126,12 @@ public class MateMemberController {
     /**
      * 디저트 메이트 멤버 신청 거절 api
      * */
+    @Operation(summary = "디저트 메이트 멤버 신청 거절", description = "디저트 메이트 멤버 신청 거절합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 멤버 신청 거절 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_MEMBER_NOT_FOUND, ErrorCode.USER_NOT_FOUND})
     @DeleteMapping("/reject")
     public ResponseEntity<Map<String, String>> rejectMemeber(@PathVariable UUID mateUuid, @RequestBody MateApplyMemberRequest request) {
 
@@ -99,6 +144,12 @@ public class MateMemberController {
     /**
      * 디저트 메이트 멤버 강퇴 api
      * */
+    @Operation(summary = "디저트 메이트 멤버 신청 강퇴", description = "디저트 메이트 멤버 신청 강퇴합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 멤버 신청 강퇴 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_MEMBER_NOT_FOUND, ErrorCode.USER_NOT_FOUND, ErrorCode.MATE_PERMISSION_DENIED})
     @DeleteMapping("/members")
     public ResponseEntity<Map<String, String>> bannedMember(@PathVariable UUID mateUuid, @RequestBody MateApplyMemberRequest request) {
 
@@ -111,6 +162,12 @@ public class MateMemberController {
     /**
      * 디저트 메이트 멤버 탈퇴 api
      * */
+    @Operation(summary = "디저트 메이트 멤버 신청 탈퇴", description = "디저트 메이트 멤버 신청 탈퇴합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트 메이트 멤버 신청 탈퇴 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_MEMBER_NOT_FOUND})
     @DeleteMapping("/leave")
     public ResponseEntity<Map<String, String>> leaveMember(@PathVariable UUID mateUuid) {
