+++ src/main/java/org/swyp/dessertbee/community/mate/controller/SavedMateController.java
@@ -1,12 +1,16 @@
 package org.swyp.dessertbee.community.mate.controller;
 
 
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.swyp.dessertbee.common.annotation.ApiErrorResponses;
 import org.swyp.dessertbee.common.exception.BusinessException;
 import org.swyp.dessertbee.common.exception.ErrorCode;
 import org.swyp.dessertbee.community.mate.dto.response.MatesPageResponse;
@@ -27,6 +31,12 @@ public class SavedMateController {
     /**
      * 디저트메이트 저장
      * */
+    @Operation(summary = "디저트메이트 저장", description = "디저트메이트 저장합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트메이트 저장 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_NOT_FOUND, ErrorCode.DUPLICATION_SAVED_MATE, ErrorCode.USER_NOT_FOUND})
     @PostMapping("/{mateUuid}")
     public ResponseEntity<Map<String, String>> saveMate(@PathVariable UUID mateUuid) {
 
@@ -42,8 +52,14 @@ public class SavedMateController {
 
 
     /**
-     * 디저트메이트 삭제
+     * 디저트메이트 저장 삭제
      * */
+    @Operation(summary = "디저트메이트 저장 삭제", description = "저장한 디저트메이트 삭제합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "디저트메이트 저장 삭제 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.MATE_NOT_FOUND, ErrorCode.SAVED_MATE_NOT_FOUND, ErrorCode.USER_NOT_FOUND})
     @DeleteMapping("/{mateUuid}")
     public ResponseEntity<Map<String, String>> deleteSavedMate(@PathVariable UUID mateUuid) {
 
@@ -58,6 +74,12 @@ public class SavedMateController {
     /**
      * 저장된 디저트메이트 조회
      * */
+    @Operation(summary = "저장된 디저트메이트 조회", description = "저장된 디저트메이트 조회합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "저장된 디저트메이트 조회 성공"),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청")
+    })
+    @ApiErrorResponses({ErrorCode.USER_NOT_FOUND, ErrorCode.INVALID_RANGE})
     @GetMapping
     private ResponseEntity<MatesPageResponse> getSavedMates(
             @RequestParam(required = false, defaultValue = "0") int from,
